% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manipulate_vafs.R
\name{compute_mutation_counts_per_sample}
\alias{compute_mutation_counts_per_sample}
\title{Compute mutation burden per sample}
\usage{
compute_mutation_counts_per_sample(
  sample_meta_path,
  variant_meta_path,
  variant_id,
  include_variants = FALSE,
  normalize_columns = FALSE
)
}
\arguments{
\item{sample_meta_path}{a file path}

\item{variant_meta_path}{a file path}

\item{variant_id}{a character vector containing one or more requested variants in the format CHROM-POS-REF-ALT}

\item{return_genotypes}{a logical value. If TRUE, returns 0 or 1 values. If FALSE, returns VAFs.}
}
\value{
a tibble where the first column is the sample identifier, and all remaining columns are the variant mutations or VAFs, except for the AC column, which is the sum of the individual variants
}
\description{
Compute mutation burden per sample
}
