% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manipulate_vafs.R
\name{subset_vaf_matrix_by_id}
\alias{subset_vaf_matrix_by_id}
\title{Extract the VAFs of a specific set of variants}
\usage{
subset_vaf_matrix_by_id(
  sample_meta_path,
  variant_meta_path,
  variant_id,
  return_genotypes = FALSE,
  normalize_columns = FALSE
)
}
\arguments{
\item{sample_meta_path}{a file path}

\item{variant_meta_path}{a file path}

\item{variant_id}{a character vector containing one or more requested variants in the format CHROM-POS-REF-ALT}

\item{return_genotypes}{a logical value. If TRUE, returns 0 or 1 values. If FALSE, returns VAFs.}
}
\value{
a tibble where the first column is the sample identifier, and all remaining columns are the variant mutations or VAFs
}
\description{
Extract the VAFs of a specific set of variants
}
